Very basic support for recording to a file.
Allows specifying a file name on the path, and this file will be
written with each output sample.  It records only a raw signed 16-bit
file, which can be converted to .wav using sox.  Doesn't do any
buffering for the writing (depends on file stream buffering), so it
can cause real-time audio glitches, but the written file will not be
missing any data.